Added unpacked game verification
The tool is now capable of verifying unpacked game files as well as supporting modifying the ammount of concurrency. The tool is now fully I/O bound which is the best case scenario for this type of tool. A lot of the code has also been refactored to make it more readable in case anyone wants to contribute (lol).